llm-service/requirements.txt
fastapi[standard]
google-genai
python-dotenv
pydantic>=2